+++ src/platform/win32/backend.rs
@@ -474,7 +474,7 @@ impl Backend {
     pub(crate) fn submit_force_unoccupied_sync(
         &self,
         vehicle: u16,
-        seat: i32,
+        seat: u8,
     ) -> Result<CommandId, DirectClientError> {
         self.state.submit_force_unoccupied_sync(vehicle, seat)
     }
